feat(store): autogenerate v3 delegate endpoints, drop custom bridge
Generate the v3 delegate endpoints from the CGW OpenAPI spec (v3 paths + DeleteDelegateV3Dto/UpdateDelegateV3Dto) instead of the hand-written bridge module.
